--- client/src/app/modules/model-manager/components/hugging-face-import-ribbon-content/huggingface-model-details/markdown/index.ts
@@ -1,7 +1,0 @@
-export interface IMarkdownParserOptions {
-  highlight: boolean;
-}
-
-export interface IMarkdownParser {
-  parse: (markdownTest: string, options?: IMarkdownParserOptions) => Promise<string>;
-}
