+++ client/src/app/shared/components/markdown-text/markdown.scss
