+++ client/src/styles.scss
@@ -16,8 +16,7 @@
 @import 'styles/global/key-val';
 @import 'styles/global/select-panel';
 
-@import 'app/modules/model-manager/components/hugging-face-import-ribbon-content/huggingface-model-details/markdown/markdown';
-// Angular Material global styles
+@import 'app/shared/components/markdown-text/markdown';
 
 @include wb-radio-group();
 @include wb-dialog();
